govseva_project/requirements.txt
Flask==3.0.3
Flask-Login==0.6.3
Flask-WTF==1.2.1
email-validator==2.1.1
WTForms==3.1.2
SQLAlchemy==2.0.32
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.1
bcrypt==4.2.0
